Create MKS_EL.cfg
Adds a seatmap so that, if users use KerbalStats, Kerbals get workshop credit for living in the workspace, Pioneer module, or orbital dock.  The workspace also becomes an EPL workshop and survey station (stats similar to the Pioneer -- it's not a specialized rocket workshop like rabidninjawombat's, but it's a better workspace than, say, the stock lab).