Add missing radian conversion
I strongly suspect this is the cause of the logistics hub NaN bug reports.

Without this fix, `a` will sometimes be negative, and so the `Sqrt(a)` on the following line will sometimes be NaN. This problem will only occur for certain destination vessel latitudes, which form a series of stripes 3.14 degrees wide and 3.14 degrees apart, starting 1.57 degrees above/below the equator. This explains the weird mix of intermittency and consistency in the reports of this bug: if the destination is landed at the wrong latitude, you're in trouble, but if it's in orbit you can just wait: every orbit spends at least half its time in a "safe" latitude, and low-inclination orbits may spend all or nearly all their time in the safe band around the equator (in particular, I bet ships placed with hyperedit virtually never see this bug).

The problem does not always occur when in an 'unsafe' latitude band. It will get less common as the source and destination get closer together in longitude, and to some extent as they get farther apart in latitude. In some cases that may provide the basis for a workaround.

The problem only occurs in surface->surface and orbit->orbit transfers, because those are the only transfers that use this function.