Fix MKS_PDU stack nodes
Found that Workspaces did not connect through the top node via CLS. After investigation it was revealed that the top and bottom stack node coordinates had been reversed on all of the MK-II MKS Modules. This fix simply swaps the node_stack_top and node_stack_bottom values in the cfg. Tested in game on the launchpad at KSC.