Specific Heat Values for TAC-LS resources
Assumptions (Food):

Food is freeze-dried for storage (I also assume greenhouses have drying facilities)
Food has high protein content (mainly because that's what's in the paper but it might also combat atrophy--assuming that high protein isn't toxic)
Food is stored at ambient temperature (CLS passable tanks)
Ambient temperature is 298 K (Comfy, climate-controlled space capsule)

Reference:
http://nopr.niscair.res.in/bitstream/123456789/7108/1/IJCT%2013(6)%20597-604.pdf (page 6)

Assumptions (Waste):

Waste is dessicated for odor and pathogen control
Waste is kept in storage but not climate-controlled storage, so it's cooler


Reference:
https://www.jstage.jst.go.jp/article/jsam/71/1/71_1_1_131/_pdf (page 2)

Assumptions (Wastewater):

The hsp is essentially equal to that of water because the impurity content is low enough to be beneath the given resolution.  Most wastewater is about 99.94 % water and .06 % impurities, and that includes the solid waste component--space wastewater is likely more pure than that.  To show it in terms of hsp would require at least five-place accuracy, and the hsp values given are four-place.

Reference:

Manual of Instruction for Sewage Treatment Plant Operators, NY State Department of Health (page 3)