Added USI_ModuleResourceWarehouse to Ranger Agr.
Added USI_ModuleResourceWarehouse to Ranger Agriculture Module, so it's not required to have a storage module in the same vessel to share resources even if they are very low. The way it is now it waits to get to 200 Supplies to start sharing, so this change would allow sharing at any time.