Remove uranium storage from Karibou
Several Karibou kontainer parts had the option to store EnrichedUranium
plus DepletedUranium, which appears to be a bug since DepletedUranium is
no longer used (it's superseded by DepletedFuel).  Furthermore, none of
the other (non-Karibou) switchable kontainer parts have the option to
store uranium; there are specialized parts for that purpose instead.

This change removes the EnrichedUranium and DepletedUranium option from
the Karibou parts that had it.