Add attachment node to MiniDrill
MKS_Drill_01 has a "back" attachment node, which is helpful for
attaching it precisely to trusses with KIS.  This change patches the
stock MiniDrill (on which MKS_Drill_01 is based) to have a similar
attachment node.

Since the stock drill is smaller than the MKS one, the attachment node
on the stock one is positioned a bit higher so that the two drills have
roughly the same ground clearance when extended.