Remove Part Deletion - Alternatives depreciated.
Since EL is being depreciated and the MKS-specific parts are currently depreciated, letting people use the default EL parts would be nice - as there are no alternatives.  (Without third-party mods, of course.)

The rest of the config works and lets people use the MKS resource chain with EL.  I say leave that and treat it like the TAC-LS config: It's shipped, and if people submit PR's it'll be updated, but no effort is made on your part to keep it up to date.