Add fore/aft attachment nodes to KER_Hub roof rack
The Karibou multi-hub has a "roof rack" with an attachment node in the
middle, but the part is long enough that it can also fit several smaller
objects (e.g. 0.625m) instead of one big object (e.g. 1.25m).  To make
that easier, I've added a pair of additional attachment nodes to both
the upper and lower levels of the roof rack, positioned ahead of and
behind the one in the middle.