Move automated drills into appropriate CTT categories
When CTT is installed, the three regular MKS drills are placed into different
tech categories, but the drills' automated variants were added later and
didn't get patched.  It makes more sense to have the automated drills in the
same tech nodes as their non-automated counterparts.  (And if any other drill
variants are added in the future, they should get the same treatment, so I've
used a wildcard to match them in the patch.)