Change functionality of cancel route to reduce route
Previously the canceling of a route was only possible when all resources at the destination were unused.
This patch changes this behavior, so that the cancel route button now reduces the route as much as possible.
This QoL change helps with managing complicated route-setups where resources are partially used by Hoppers somewhere in a sequential transfer route.

This is done by calculating the maximal reducible amount of resources and using this number for the cancellation.